rename dateTime to dateTimePicker and slider to rangeSlider
